refactor: Lottos 역할 일부 Purchaser로 이전
